Improve role based on review
First, change the way to check if an installation exists.
Now the build is always done under /opt and is no cleared.
Then the deployment is just a symlink to there.

Thanks to this a lot of the complexity of main.yml could be removed.

Also fix minor mistakes pointed out in review.